Update fund-sets.tex
In the Core Ideas, there's a contradiction in the explanation of what a function is with what is given as requirement for a function in the main text. I propose to change 'unique' to 'single', since the object in the second collection aka the range can be assigned to twice. Only the domain objects have to be unique, if I interprete the text correctly....